Refactor task specific modals location
